requirements.txt
# API
fastapi==0.104.1
uvicorn[standard]==0.23.2
pydantic==1.10.8
python-multipart==0.0.6
python-dotenv==1.0.0

# Speech-to-text
openai==1.1.1
google-cloud-speech==2.21.0

# Data extraction
python-dateutil==2.8.2

# Google Sheets integration
google-api-python-client==2.106.0
google-auth==2.23.3
